+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## 4.0.0
 
-- Add font size setting for editor and terminal.
-- Add font family setting for terminal.
+- Add commands to change editor and terminal font sizes.
+- Add command to change terminal font family.
 - Add ability to add and remove fonts.
 - Remove `font-switcher.fontFamily`  and `font-switcher.enableLivePreview` settings.
   - Live preview is always enabled and the extension uses the default `editor.fontFamily` setting.